+++ .vs/VSWorkspaceState.json
@@ -1,7 +1,3 @@
 {
-  "ExpandedNodes": [
-    ""
-  ],
-  "SelectedNode": "\\gpio.cpp",
   "PreviewInSolutionExplorer": false
 }
